Add '_all' shorthand for all entity types
